combine three config files
